src/requirements.txt
opencc-python-reimplemented
bs4